requirements.txt
python-dotenv
requests
pandas
sqlalchemy
psycopg2-binary
geopandas
shapely
scikit-learn
statsmodels
xgboost
lightgbm
joblib
streamlit
folium
plotly
optuna
fastapi
uvicorn[standard]
pydantic
celery
redis
